refactor int folding to separate repo
